+++ assets-src/admin.js
@@ -65,7 +65,8 @@ export default class ACFML {
    */
   initTranslatablePostTitle() {
     acf.addAction(`ready_field/key=field_acfml_post_title_${acfml.defaultLanguage}`, $field => {
-      $field.$input().attr('id', 'title');
+      $('#titlediv').remove();
+      // $field.$input().attr('id', 'title');
       if( !acfml.isMobile && !$field.val() ) $field.$input().focus();
     });
     acf.addAction(`ready_field/key=field_acfml_slug`, $field => {