+++ changelog.md
@@ -1,5 +1,6 @@
 #### 0.0.1 (2020-12-16)
 
+- 'request' for translated slugs 🎉 (#53a7aea)
 - First draft of unique post slug (#1cbaedb)
 - Add slugs (#e086b4b)
 - Better link filters (#4643582)